+++ src/components/KSRSForm/classes.ts
@@ -1,0 +1,6 @@
+export const labelClass = `font-bold leading-4 content-center text-left`;
+export const labelRequirementClass = 'text-sm font-normal text-grey-400';
+export const labelRequiredClass = 'text-red-600';
+export const checkBoxClass = 'accent-primary-600 focus:accent-primary-600';
+export const formFieldClass = `rounded-lg bg-grey-100 p-4 px-4 grow flex flex-col gap-2 gap-y-3 dark:bg-grey-700`;
+export const fieldClass = `rounded-lg p-3 py-3 shadow-md text-grey-600 dark:bg-grey-900 dark:shadow-none dark:text-grey-200 placeholder:text-grey-300 dark:placeholder:text-grey-400 focus:outline-2 focus:outline-primary-600 focus:ring-inset focus:outline-none focus:shadow-none focus:outline-offset-0 `;
