+++ src/components/KSRSForm/index.tsx
@@ -1,8 +1,10 @@
 'use client';
 import { useState, useEffect } from 'react';
 import { ServiceSummary, ProductivityData } from '@/types';
-import { convertToHHMM, convertToMinutesSeconds } from '@/utilities/timeConverter';
-import { createSummary } from './summary';
+import { parseServiceSummaryData } from './functions/parseServiceSummary';
+import { parseProductivityData } from './functions/parseProductivity';
+import { Checkbox } from './components/Checkbox';
+import { TextArea } from './components/Textarea';
 
 type KSRSFormProps = {
   onSubmit: (
@@ -102,197 +104,6 @@ const KSRSForm = ({ onSubmit, initialValues }: KSRSFormProps) => {
     const value = event.target.value;
     setCopiedProdData(value);
     setProdData(value);
-  };
-
-  const parseServiceSummaryData = (data: string): ServiceSummary => {
-    const lines = data
-      .split('\n')
-      .map((line) => line.trim())
-      .filter((line) => line !== '');
-    const firstLine = lines[0];
-    const firstPeriodIndex = firstLine.indexOf('.');
-    const siteName = firstLine.substring(0, firstPeriodIndex).trim();
-    const cleanSiteName = siteName.replace(/\*\*(.*?)\*\*/, '$1').trim();
-    const range = firstLine.substring(firstPeriodIndex + 1).trim();
-
-    const site = cleanSiteName || firstLine.trim();
-    lines.shift();
-
-    const summary = createSummary(site, range);
-
-    for (const line of lines) {
-      const [key, ...values] = line.split('\t').map((v) => v.trim());
-
-      switch (key) {
-        case 'Average Delivery Time':
-          const Delivery = summary.averageDeliveryTime;
-
-          Delivery.starters = convertToMinutesSeconds(parseFloat(values[0]));
-          Delivery.mains = convertToMinutesSeconds(parseFloat(values[1]));
-          Delivery.desserts = convertToMinutesSeconds(parseFloat(values[2]));
-          Delivery.total = convertToMinutesSeconds(parseFloat(values[3]));
-          break;
-
-        case 'Average Wait Time':
-          const Wait = summary.averageWaitTime;
-
-          Wait.starters = convertToMinutesSeconds(parseFloat(values[0]));
-          Wait.mains = convertToMinutesSeconds(parseFloat(values[1]));
-          Wait.desserts = convertToMinutesSeconds(parseFloat(values[2]));
-          Wait.total = convertToMinutesSeconds(parseFloat(values[3]));
-          break;
-
-        case 'Average Preparation Time':
-          const Prep = summary.averagePreparationTime;
-
-          Prep.starters = convertToMinutesSeconds(parseFloat(values[0]));
-          Prep.mains = convertToMinutesSeconds(parseFloat(values[1]));
-          Prep.desserts = convertToMinutesSeconds(parseFloat(values[2]));
-          Prep.total = convertToMinutesSeconds(parseFloat(values[3]));
-          break;
-
-        case 'No. of Orders':
-          summary.numberOfOrders = parseInt(values[3], 10);
-          break;
-
-        case 'No. of Late Orders':
-          const Orders = summary.numberOfOrders;
-
-          const Starters = summary.numberOfLateOrders.starters;
-          Starters.count = parseInt(values[0], 10);
-          Starters.percentage = Math.round((Starters.count / Orders) * 100);
-
-          const Mains = summary.numberOfLateOrders.mains;
-          Mains.count = parseInt(values[1], 10);
-          Mains.percentage = Math.round((Mains.count / Orders) * 100);
-
-          const Desserts = summary.numberOfLateOrders.desserts;
-          Desserts.count = parseInt(values[2], 10);
-          Desserts.percentage = Math.round((Desserts.count / Orders) * 100);
-
-          summary.numberOfLateOrders.total.count = parseInt(values[3], 10);
-          summary.numberOfLateOrders.total.percentage = Math.round(
-            (summary.numberOfLateOrders.total.count / Orders) * 100,
-          );
-          break;
-
-        case 'No. of Items':
-          summary.numberOfItems = parseInt(values[3], 10);
-          break;
-
-        case 'No. of Late Items':
-          const Items = summary.numberOfItems;
-
-          const StarterItems = summary.numberOfLateItems.starters;
-          StarterItems.count = parseInt(values[0], 10);
-          StarterItems.percentage = Math.round((StarterItems.count / Items) * 100);
-
-          const MainsItems = summary.numberOfLateItems.mains;
-          MainsItems.count = parseInt(values[1], 10);
-          MainsItems.percentage = Math.round((MainsItems.count / Items) * 100);
-
-          const DessertItems = summary.numberOfLateItems.desserts;
-          DessertItems.count = parseInt(values[2], 10);
-          DessertItems.percentage = Math.round((DessertItems.count / Items) * 100);
-
-          summary.numberOfLateItems.total.count = parseInt(values[3], 10);
-          summary.numberOfLateItems.total.percentage = Math.round(
-            (summary.numberOfLateItems.total.count / Items) * 100,
-          );
-
-          break;
-
-        case 'Table/Meal Checks On-Time':
-          summary.checksOnTime.onTime = parseInt(values[0], 10);
-          summary.checksOnTime.early = parseInt(values[1], 10);
-          summary.checksOnTime.late = parseInt(values[2], 10);
-          break;
-
-        case 'CHEF1':
-          const Chef = summary.chef1;
-
-          Chef.averagePrepTime = convertToMinutesSeconds(parseFloat(values[0]));
-          Chef.numberOfOrders = parseInt(values[1], 10);
-          Chef.ordersLate.count = parseInt(values[2], 10);
-          Chef.ordersLate.percentage = Math.round(
-            (Chef.ordersLate.count / Chef.numberOfOrders) * 100,
-          );
-
-          Chef.numberOfItems = parseInt(values[3], 10);
-          Chef.itemsLate.count = parseInt(values[4], 10);
-          Chef.itemsLate.percentage = Math.round((Chef.itemsLate.count / Chef.numberOfItems) * 100);
-
-          Chef.ordersBumped = parseInt(values[5], 10);
-          Chef.manualHolds = parseInt(values[6], 10);
-          break;
-
-        case 'DISPENSE':
-          const Dispense = summary.dispense;
-
-          Dispense.averagePrepTime = convertToMinutesSeconds(parseFloat(values[0]));
-          Dispense.numberOfOrders = parseInt(values[1], 10);
-          Dispense.ordersLate.count = parseInt(values[2], 10);
-          Dispense.ordersLate.percentage = Math.round(
-            (Dispense.ordersLate.count / Dispense.numberOfOrders) * 100,
-          );
-          Dispense.numberOfItems = parseInt(values[3], 10);
-          Dispense.itemsLate.count = parseInt(values[4], 10);
-          Dispense.itemsLate.percentage = Math.round(
-            (Dispense.itemsLate.count / Dispense.numberOfItems) * 100,
-          );
-          Dispense.ordersBumped = parseInt(values[5], 10);
-          Dispense.manualHolds = parseInt(values[6], 10);
-          break;
-
-        default:
-          break;
-      }
-    }
-
-    return summary;
-  };
-
-  const parseProductivityData = (data: string): ProductivityData => {
-    const lines = data
-      .split('\n')
-      .map((line) => line.trim())
-      .filter((line) => line !== '');
-
-    const firstLine = lines[0];
-    const firstPeriodIndex = firstLine.indexOf('.');
-    const range = firstLine.substring(firstPeriodIndex + 1).trim();
-
-    lines.shift();
-    lines.shift();
-
-    const productivity: ProductivityData = {
-      range: range,
-      staffMembers: [],
-    };
-
-    for (const line of lines) {
-      const [station, name, avgTime, noOrders, noItems, ordersLate, longestOrder, hoursWorked] =
-        line.split('\t').map((v) => v.trim());
-
-      if (station && name) {
-        const lateOrdersCount = parseInt(ordersLate, 10);
-        const totalOrders = parseInt(noOrders, 10);
-
-        productivity.staffMembers.push({
-          name,
-          prepTime: convertToMinutesSeconds(avgTime),
-          orders: totalOrders,
-          items: parseInt(noItems, 10),
-          lateOrders: lateOrdersCount,
-          lateOrdersPercentage:
-            totalOrders > 0 ? Math.round((lateOrdersCount / totalOrders) * 100) : 0,
-          longestOrder: convertToMinutesSeconds(longestOrder),
-          hoursWorked: convertToHHMM(hoursWorked),
-        });
-      }
-    }
-
-    return productivity;
   };
 
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
@@ -395,81 +206,55 @@ const KSRSForm = ({ onSubmit, initialValues }: KSRSFormProps) => {
         </div>
 
         <div className={`flex flex-col gap-4 flex-wrap w-full md:flex-row md:col-span-4 `}>
-          <div className={`${formFieldClass} !flex-row gap-x-3 md:justify-center`}>
-            <label htmlFor='kitchenLates' className={`${labelClass}`}>
-              Kitchen Lates: <span className='text-sm font-normal text-grey-400'>(Optional)</span>
-            </label>
-            <input
-              type='checkbox'
-              id='kitchenLates'
-              checked={kitLates}
-              onChange={handleKitLateChange}
-              className='accent-primary-600 focus:accent-primary-600'
-            />
-          </div>
+          <Checkbox
+            id='kitchenLates'
+            label='Kitchen Lates'
+            checked={kitLates}
+            onChange={handleKitLateChange}
+            required={false}
+            containerClassName={`${formFieldClass} !flex-row gap-x-3 md:justify-center`}
+          />
 
-          <div className={`${formFieldClass} !flex-row gap-x-3 md:justify-center`}>
-            <label htmlFor='floorLates' className={`${labelClass}`}>
-              Floor Lates: <span className='text-sm font-normal text-grey-400'>(Optional)</span>
-            </label>
-            <input
-              type='checkbox'
-              id='floorLates'
-              checked={floorLates}
-              onChange={handleFloorLateChange}
-              className='accent-primary-600 focus:accent-primary-600'
-            />
-          </div>
+          <Checkbox
+            id='floorLates'
+            label='Floor Lates'
+            checked={floorLates}
+            onChange={handleFloorLateChange}
+            required={false}
+            containerClassName={`${formFieldClass} !flex-row gap-x-3 md:justify-center`}
+          />
 
-          <div className={`${formFieldClass} !flex-row gap-x-3 md:justify-center`}>
-            <label htmlFor='foodLift' className={`${labelClass}`}>
-              Food Lift: <span className='text-sm font-normal text-grey-400'>(Optional)</span>
-            </label>
-            <input
-              type='checkbox'
-              id='foodLift'
-              checked={lift}
-              onChange={handleLiftChange}
-              className='accent-primary-600 focus:accent-primary-600'
-            />
-          </div>
+          <Checkbox
+            id='foodLift'
+            label='Food Lift'
+            checked={lift}
+            onChange={handleLiftChange}
+            required={false}
+            containerClassName={`${formFieldClass} !flex-row gap-x-3 md:justify-center`}
+          />
         </div>
 
-        <div className={`${formFieldClass} col-span-4`}>
-          <label htmlFor='serviceSummaryData' className={`${labelClass}`}>
-            Service Summary Data:{' '}
-            <span className='text-sm font-normal text-grey-400'>
-              (Required<span className='text-red-600'>*</span>)
-            </span>
-          </label>
-          <textarea
-            id='serviceSummaryData'
-            value={serviceData}
-            onChange={handleServiceChange}
-            rows={13}
-            placeholder='Copy and paste the service summary report here'
-            className={`${fieldClass}`}
-            required
-          ></textarea>
-        </div>
+        <TextArea
+          id='serviceSummaryData'
+          label='Service Summary Data'
+          value={serviceData}
+          onChange={handleServiceChange}
+          rows={13}
+          placeholder='Copy and paste the service summary report here'
+          containerClassName={`${formFieldClass} col-span-4`}
+          required
+        />
 
-        <div className={`${formFieldClass} col-span-4`}>
-          <label htmlFor='prodData' className={`${labelClass}`}>
-            Productivity Data:{' '}
-            <span className='text-sm font-normal text-grey-400'>
-              (Required<span className='text-red-600'>*</span>)
-            </span>
-          </label>
-          <textarea
-            id='prodData'
-            value={prodData}
-            onChange={handleProdChange}
-            placeholder='Copy and paste productivity report here'
-            rows={6}
-            className={`${fieldClass}`}
-            required
-          ></textarea>
-        </div>
+        <TextArea
+          id='prodData'
+          label='Productivity Data'
+          value={prodData}
+          onChange={handleProdChange}
+          rows={6}
+          placeholder='Copy and paste productivity report here'
+          containerClassName={`${formFieldClass} col-span-4`}
+          required
+        />
 
         <button
           type='submit'
